+++ src/bigdict/__init__.py
@@ -2,7 +2,7 @@
 Bigdict
 """
 
-__version__ = '0.3.3'
+__version__ = '0.3.4'
 
 from ._bigdict import Bigdict, ReadonlyError